requirements.txt
beautifulsoup4
requests
telebot
telegram
python-dotenv